Fixing and adding missing features to position_sub
These are the changes that I've made to position_sub.py. I fixed some errors in the code, but I'm not sure if there's other errors that I've made, as I'm not very familiar with Python. I fixed the code that had been commented out, and fixed the getCenterCommand, which was incomplete and wrong as far as I could see. The setpoints did not match the errors. For example, in the original code, when y is greater than yError, the setpoint appends "go left," but this isn't correct at all. It should say "go down," because the y coordinate is either up or down, and when the y is greater than the yError, that means that the AUV should go down. Again, not sure if this code will actually run or compile, but these are just my adjustments. Testing the code is a little harder, so haven't been able to do that, but let me know how this code is.