Update execute_gate.py
This is an updated execute_gate file, but it's still incomplete, as I don't know yet what MoveForwardNoVision is supposed to do. Like all the other code, it's still untested. I'll probably test it some time this week, so there's probably still errors, but this is what I've done so far.